CR comments 2
