+++ cmd/authproxy/app/authproxy.go
@@ -17,15 +17,11 @@ limitations under the License.
 package app
 
 import (
-	"fmt"
-	"net/http"
-
 	"github.com/nuclio/nuclio/pkg/auth"
 	"github.com/nuclio/nuclio/pkg/loggersink"
 	"github.com/nuclio/nuclio/pkg/platformconfig"
 
 	"github.com/nuclio/errors"
-	"github.com/nuclio/logger"
 
 	// register logger sinks (stdout, appinsights) into the loggersink registry via init()
 	_ "github.com/nuclio/nuclio/pkg/sinks"
@@ -60,87 +56,6 @@ func Run(mode auth.ProxyMode,
 		return errors.Wrap(err, "Failed to start auth-proxy server")
 	}
 	select {}
-}
-
-// server fronts a function or the DLX.
-type server struct {
-	logger      logger.Logger
-	httpServer  *http.Server
-	upstreamURL string
-	authURL     string
-	signinURL   string
-}
-
-// newServer builds the auth-proxy for the given mode. The mode decides only where it listens and what it
-// serves: reverseProxy is exposed to the cluster on a configurable port and forwards to the processor;
-// authOnly is reachable only from within the pod (loopback) and serves the DLX /auth endpoint.
-func newServer(parentLogger logger.Logger,
-	mode auth.ProxyMode,
-	listenPort int,
-	upstreamURL string,
-	authURL string,
-	signinURL string) (*server, error) {
-
-	var listenAddress string
-	var handler http.Handler
-	authLogger := parentLogger.GetChild("authproxy")
-
-	switch mode {
-	case auth.ProxyModeReverseProxy:
-		listenAddress = fmt.Sprintf(":%d", listenPort)
-		handler = newReverseProxyHandler(authLogger)
-	case auth.ProxyModeAuthOnly:
-		listenAddress = fmt.Sprintf("127.0.0.1:%d", listenPort)
-		handler = newAuthOnlyHandler(authLogger)
-	default:
-		return nil, errors.Errorf("Unknown auth-proxy mode: %s", mode)
-	}
-
-	return &server{
-		logger: authLogger,
-		httpServer: &http.Server{
-			Addr:    listenAddress,
-			Handler: handler,
-		},
-		upstreamURL: upstreamURL,
-		authURL:     authURL,
-		signinURL:   signinURL,
-	}, nil
-}
-
-func (s *server) start() error {
-	s.logger.InfoWith("Starting auth-proxy",
-		"listenAddress", s.httpServer.Addr,
-		"upstreamURL", s.upstreamURL,
-		"authURL", s.authURL,
-		"signinURL", s.signinURL)
-
-	if err := s.httpServer.ListenAndServe(); err != nil {
-		return errors.Wrap(err, "Auth-proxy server failed to start")
-	}
-
-	return nil
-}
-
-// newReverseProxyHandler serves the running-function topology. It does not forward requests upstream yet:
-// authenticating and proxying to the processor over loopback is implemented in NUC-828 / NUC-837.
-func newReverseProxyHandler(logger logger.Logger) http.Handler {
-	return http.HandlerFunc(func(responseWriter http.ResponseWriter, _ *http.Request) {
-		logger.Warn("reverseProxy forwarding is not yet implemented")
-		responseWriter.WriteHeader(http.StatusNotImplemented)
-	})
-}
-
-// newAuthOnlyHandler serves the DLX /auth decision endpoint. Only /auth is routed; any other path
-// returns 404, pinning the endpoint surface now so the NUC-837 decision logic slots into /auth.
-func newAuthOnlyHandler(logger logger.Logger) http.Handler {
-	mux := http.NewServeMux()
-	mux.HandleFunc("/auth", func(responseWriter http.ResponseWriter, _ *http.Request) {
-		//TODO: implement authOnly decision as part of NUC-837
-		logger.Warn("authOnly mode was requested but is not yet implemented")
-		responseWriter.WriteHeader(http.StatusNotImplemented)
-	})
-	return mux
 }
 
 func validateConfiguration(listenPort int, upstreamURL string, authURL string, signinURL string) error {
